restructure project into src
