+++ src/components/IdeaList.tsx
@@ -1,7 +1,7 @@
 // components/IdeaList.tsx (Project File View)
 import React, { useState, useRef, ChangeEvent, useCallback } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
-import { Project, Idea, Attachment } from '../types';
+import { Project, Idea, Attachment } from '../../types';
 import Button from './Button';
 import Modal from './Modal';
 import * as fileService from '../services/fileService';
@@ -12,7 +12,7 @@ import {
     TrashIcon, PhotoIcon, ArrowUpTrayIcon, DocumentIcon, ArrowDownTrayIcon, PaperClipIcon,
     DocumentChartBarIcon // New icon
 } from './icons'; 
-import { staggerContainer, staggerItem } from '../motion/variants';
+import { staggerContainer, staggerItem } from '../../motion/variants';
 import { useSortedIdeas } from '../utils/zineExporter';
 
 interface IdeaListProps {
