+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './src/index.css';
-import App from './App';
+import './index.css';
+import App from './pages/App';
 import ErrorBoundary from './components/ErrorBoundary'; // Import the ErrorBoundary
 
 const rootElement = document.getElementById('root');
