+++ tsconfig.json
@@ -24,7 +24,7 @@
     "noUncheckedSideEffectImports": true,
 
     "paths": {
-      "@/*" :  ["./*"]
+      "@/*": ["./src/*"]
     }
   },
   "exclude": ["**/*.test.ts", "**/*.test.tsx", "node_modules", "coverage"]
